+++ src/types.d.ts
@@ -1,2 +1,1 @@
 declare module "markdown-it-task-lists";
-declare module "markdown-it-katex";